Fixed frame counter width.
Only used 8bit vars, while 16bit were required.

+++ CubeFirmware/header/generator.h
@@ -21,5 +21,5 @@
 
 // Generate 8 frames, store them from index start.
 // data has 8 bytes --> 1 layer. This layer is moved from back to front.
-void generateMovingAnimation(uint8_t *data, uint8_t start, uint8_t duration);
-void renderText(char *text, uint8_t start);
+void generateMovingAnimation(uint8_t *data, uint16_t start, uint8_t duration);
+void renderText(char *text, uint16_t start);
