+++ CubeFirmware/header/transmit.h
@@ -28,7 +28,7 @@ void sendAudioData(void);
 void printTime(void);
 void dumpFrame(uint8_t *f);
 uint8_t *readLine(void);
-uint8_t readNumber(uint8_t base);
+uint16_t readNumber(uint8_t base);
 void writeNumber(uint8_t num, uint8_t base);
 uint8_t *readAFrame(void);
 void simpleAnimationInput(void);
